ai_backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pillow==11.3.0
numpy==2.3.2

# Optional open-source VLM support:
# Install these only if you want mode=vlm. They are intentionally optional
# because model downloads and runtime requirements are larger.
# transformers==4.41.2
# torch==2.3.1
